requirements.txt
openai~=1.88.0
yookassa~=3.5.0
pillow~=11.2.1
aiogram~=3.20.0.post0
python-dotenv~=1.1.0
loguru~=0.7.3
apscheduler~=3.11.0
tzlocal~=5.3.1
python-dateutil~=2.9.0.post0
asyncpg~=0.30.0
dropbox~=12.0.2